chore: bump build to 38 + watch targets release signing
- pubspec 0.1.5+38
- 워치 타깃 Release/Profile 수동 서명(fmk watch appstore / fmk watch widgets appstore), ExportOptions 에 프로파일 추가
- 워치 타깃 SUPPORTED_PLATFORMS=watchos watchsimulator (프로젝트 Release 의 iphoneos 상속으로 아카이브 시 iOS SDK 로 컴파일되던 문제)
- 워치 실기기 arm64_32(32비트 Int) 대응: 색상 리터럴 UInt32 경유

+++ ios/FmkWidgets/FmkPayloadStore.swift
@@ -286,6 +286,8 @@ extension FmkPayload {
 }
 
 /// 0(미저장/투명)이면 레드 폴백 — Kotlin accentColor()/rowColor() 와 동일 규칙.
-private func normalizedColor(_ stored: Int, fallback: Int = 0xFFEF4444) -> Int {
-  stored == 0 ? Int(Int32(bitPattern: UInt32(fallback))) : stored
+/// fallback 은 UInt32 — 워치 실기기(arm64_32)는 Int 가 32비트라 0xFF… 리터럴이
+/// Int 파라미터로 들어가면 오버플로 컴파일 에러가 난다.
+private func normalizedColor(_ stored: Int, fallback: UInt32 = 0xFFEF4444) -> Int {
+  stored == 0 ? Int(Int32(bitPattern: fallback)) : stored
 }
